refactor(bootstrap): split phase7 listener runtime
