+++ docs/backlog.md
@@ -111,6 +111,7 @@
 | PR237-T1 | **DUAL-LISTENER-TEST-GAPS-01** (Cx1, 🟡 可延后): PR-A14a 第二轮六角色 review 测试维度 10 条 Cx1 gap 合并登记：(T-01) FinalizeAuth 排序不变量——duplicate 与 Delegated 不一致交叉用例；(T-02) `shutdownBothServers` 部分错误聚合未测（只有 clean path）；(T-03) httpErrCh 双关闭在 `-race` 下未覆盖；(T-04) 测试仅覆盖 `owned=false` primary-listener 路径，`owned=true` 分支零覆盖；(T-05) `WithInternalMiddleware` 多个 mw 的顺序保真；(T-06) `FinalizeAuth` 二次调用的错误文本；(T-07) 有 metas 未 FinalizeAuth 的 `ServeHTTP` panic 文本；(T-08) 确认 `auth_integration_test.go` 的 `//go:build integration` tag 存在；(T-09) `TestAuthWiring_InternalGuard_*` 对 internal listener 的 wait-for-ready；(T-10) `runtime.NumGoroutine` baseline 捕获时机偏早。 | 3h | `runtime/http/router/*_test.go` + `runtime/bootstrap/*_test.go` + `cmd/corebundle/auth_integration_test.go` | PR #237 第二轮 reviewer (testing 维度) |
 | PR237-OB2 | **INTERNAL-LISTENER-METRICS-01** (Cx2, 🟡 可延后): PR-A14a 的 internal mux 未挂 `middleware.Metrics`——`http_requests_total` / `http_request_duration_seconds` 完全不记录 /internal/v1/* 流量。方案：(a) 给 collector 加 `listener` 标签，两个 mux 用同一 Provider 不同 label；或 (b) 新增 router.WithInternalMetricsCollector option + 第二个 collector 实例。运营可见度缺失非 P0，且 Provider 接口改造属于更大议题，暂延 PR-A14b 或单开独立 PR。 | 4h | `runtime/http/router/router.go` + `runtime/http/middleware/metrics.go` + `runtime/bootstrap/bootstrap_phases.go::autoWireHTTPMetricsCollector` | PR #237 第二轮 reviewer OB-F2 |
 | PR237-DX1 | **LISTENER-OPTION-NAMING-UNIFY-01** (Cx2, 🟡 可延后): PR-A14a 选项前缀不对称——`WithHTTPPrimaryAddr` / `WithHTTPInternalAddr`（带 HTTP）与 `WithPrimaryListener` / `WithInternalListener`（不带 HTTP）。IDE 补全时两组选项不相邻。方案：统一为 `WithHTTP*Listener` 或去掉 addr 侧的 HTTP 前缀。当前只有 gocell 自身无外部调用方，可在任意时间一次改名。 | 2h | `runtime/bootstrap/bootstrap.go` + 所有测试 + `cmd/corebundle` + `examples/*/main.go` | PR #237 第二轮 reviewer DX F3 |
+| LISTENER-API-SPEC-01 | **LISTENER-SPEC-PUBLIC-API-EVAL-01** (Cx2, 🟡 可延后, A53 明确排除): 评估将 `bootstrap.WithListener(ref, addr, authChain, opts...)` 替换为显式 `ListenerSpec` 风格 public API，使 pre-bound listener 不再需要传日志用 addr，并把 TLS / socket ownership / shutdownGrace 字段结构化表达。预计范围为 listener public API 重塑 + 调用点迁移，约 12-18h。**非目标**：在没有真实第二 transport 前不引入 `ServerGroup` / 多 transport 抽象；A53 仅做 phase7 物理拆分与 shutdown hardening，不迁移 public API。 | 12-18h | `runtime/bootstrap/listener.go` + `cmd/corebundle` + `examples/*/main.go` + bootstrap tests | A53 方案确认 |
 | ~~PR237-A2~~ | ~~**ROUTEDECL-DELEGATED-RENAME-01**~~ ✅ **DONE @ PR #278（PR-CFG-E）**：`kernel/cell.AuthRouteMeta` 已移除 `Delegated` 语义并切到 `IsInternal()` + `InternalPathPrefix` 结构化判定，`runtime/http/router` 侧一致性校验也已同步。 | — | `kernel/cell/registrar.go` + `runtime/http/router/router.go` + `cells/accesscore/slices/rbacassign/handler.go` | PR #278（commit `50cf9fde`） |
 | PR237-A4 | **INTERNAL-PREFIX-CONSTANT-RELOCATE-01** (Cx2, 🟡 可延后): `runtime/http/router/router.go::internalPathPrefix = "/internal/v1/"` 是路由包私有常量，但 `.claude/rules/gocell/api-versioning.md` 将 `/internal/v1/` 作为跨 runtime 的控制面前缀。若后续出现 v2-internal，router 包不是正确位置。建议迁到 `pkg/httputil/prefix.go` 或类似位置，与 `PublicAPIV1Prefix` 一起作为单一来源。 | 1h | `pkg/httputil/prefix.go`（新） + `runtime/http/router/router.go` | PR #237 第二轮 reviewer architect A4 |
 | PR240-A1 | **MANAGEDRESOURCE-CONTRACT-FORMALIZE-01** (Cx2, 🟡 可延后): PR-A18 把 RabbitMQ Connection 接入 `lifecycle.ManagedResource`，叠加 Vault / Postgres / Redis 已经接入，`ManagedResource` 已成事实上的 adapter 生命周期契约。建议：(a) `kernel/lifecycle/doc.go` 写实施清单（Worker() != nil 何时合理 / Checkers 命名后缀 / Close 排干约定）；(b) 加 archtest 强制 `adapters/*` 顶层 exported 类型实现 `ManagedResource` 或显式 opt out。 | 4h | `kernel/lifecycle/doc.go`（新） + `kernel/governance/rules_arch.go` | PR #240 L4 architect 跨切面 finding 1 |
