+++ runtime/bootstrap/bootstrap_phases.go
@@ -9,17 +9,13 @@ package bootstrap
 
 import (
 	"context"
-	"crypto/tls"
 	"errors"
 	"fmt"
 	"io"
 	"log/slog"
-	"net"
-	"net/http"
 	"reflect"
 	"sort"
 	"strings"
-	"sync/atomic"
 	"time"
 
 	"github.com/ghbvf/gocell/kernel/assembly"
@@ -1107,273 +1103,6 @@ func (b *Bootstrap) checkNoEventRegistrars(asm *assembly.CoreAssembly) error {
 		}
 	}
 	return nil
-}
-
-// phase7StartHTTPServer creates and starts N http.Servers — one per declared
-// listener — pre-binding sockets synchronously so port conflicts surface
-// before any goroutine is started. Bind failures for listeners after the
-// first are rolled back by closing already-owned sockets. Each server runs
-// in its own goroutine; errors fan into a single buffered channel read by
-// phase9. The LIFO teardown drains all servers in parallel under the shared
-// shutCtx budget.
-//
-// PR-A14b: replaces the hard-coded primary+internal pair with a generic
-// N-server loop driven by b.listenerConfigs.
-//
-// ref: go-kratos/kratos app.go@main L95-122 — per-server goroutine pair.
-// ref: kubernetes/apiserver pkg/server/secure_serving.go — pre-bind listener.
-// shutdownCtxFor derives the per-server shutdown context from the parent ctx
-// and the listener's shutGrace setting.
-//
-//   - shutGrace > 0 wraps the parent with context.WithTimeout(parent, shutGrace).
-//     context.WithTimeout already bounds the resulting deadline to whichever of
-//     parent.Deadline() and (now + shutGrace) comes first, so the global
-//     shutdownTimeout always wins when shutGrace exceeds it (R2-03).
-//   - shutGrace == 0 returns the parent unchanged (no per-listener override; the
-//     server inherits the global shutdownTimeout). The returned cancel is a
-//     no-op so callers can defer it unconditionally.
-func shutdownCtxFor(parent context.Context, shutGrace time.Duration) (context.Context, context.CancelFunc) {
-	if shutGrace > 0 {
-		return context.WithTimeout(parent, shutGrace)
-	}
-	return parent, noopShutdownCancel
-}
-
-// noopShutdownCancel is the cancel function returned by shutdownCtxFor when
-// shutGrace == 0 — there is nothing to cancel because we did not derive a new
-// context, but callers always defer the returned cancel so we hand back a
-// no-op rather than a nil that would NPE.
-func noopShutdownCancel() {}
-
-// boundServer holds a resolved HTTP server and its associated listener.
-type boundServer struct {
-	name      string
-	srv       *http.Server
-	ln        net.Listener
-	owned     bool          // true when bootstrap bound the socket (not caller-injected)
-	shutGrace time.Duration // 0 means inherit the global shutdownTimeout
-	authDesc  string        // OPS-09: auth chain description for startup log
-}
-
-func (b *Bootstrap) phase7StartHTTPServer(s *phaseState) error {
-	servers, err := b.phase7BindListeners(s)
-	if err != nil {
-		return err
-	}
-	if len(servers) == 0 {
-		return nil
-	}
-	httpErrCh := b.phase7ServeAll(servers)
-	s.httpErrCh = httpErrCh
-	// Convert to shutdownTasks for teardown (OPS-01: named logging via shutdownTask.name).
-	shutTasks := boundServersToTasks(servers)
-	s.addTeardown(func(c context.Context) error {
-		return shutdownAllServers(c, shutTasks)
-	})
-	return nil
-}
-
-// phase7BindListeners pre-binds all declared listeners synchronously. If any
-// bind fails, already-owned sockets are closed before returning the error.
-//
-// CORR-06: sort by ref.String() so bind order is deterministic across runs
-// and log lines appear in a consistent order for operators.
-// SEC-11: http.Server is constructed with ReadTimeout, WriteTimeout, and
-// IdleTimeout to prevent Slowloris / slow-write DoS attacks.
-// OPS-06: emit slog.Info after each successful bind (listener + addr + auth).
-// OPS-07: emit slog.Warn when a non-loopback listener binds with AuthNone or empty auth chain.
-func (b *Bootstrap) phase7BindListeners(s *phaseState) ([]boundServer, error) {
-	// Collect and sort refs for deterministic iteration order.
-	refs := make([]cell.ListenerRef, 0, len(b.listenerConfigs))
-	for ref := range b.listenerConfigs {
-		refs = append(refs, ref)
-	}
-	sort.Slice(refs, func(i, j int) bool { return refs[i].String() < refs[j].String() })
-
-	var servers []boundServer
-	for _, ref := range refs {
-		cfg := b.listenerConfigs[ref]
-		rtr, ok := s.routers[ref]
-		if !ok {
-			return nil, fmt.Errorf("bootstrap: no router for listener %q", ref.String())
-		}
-		ln, owned, err := resolveListener(cfg)
-		if err != nil {
-			closeOwnedSockets(servers)
-			slog.Error("bootstrap: failed to bind HTTP listener",
-				slog.String("listener", ref.String()),
-				slog.String("addr", cfg.addr),
-				slog.Any("error", err))
-			return nil, fmt.Errorf("bootstrap: listen %s %s: %w", ref.String(), cfg.addr, err)
-		}
-
-		authDesc := describeAuthChain(cfg.authChain)
-		slog.Info("bootstrap: HTTP listener bound",
-			slog.String("listener", ref.String()),
-			slog.String("addr", ln.Addr().String()),
-			slog.String("auth", authDesc))
-
-		// OPS-07 / F7 round-3: warn when a non-loopback address is served with
-		// AuthNone or an empty auth chain. The dangerous case is the wildcard bind
-		// (0.0.0.0 / ::): the listener is reachable on every interface, including
-		// externally routable addresses, but ServeHTTP runs without an auth gate.
-		// explicit_auth_none=true means the caller deliberately passed AuthNone{};
-		// false means the chain was nil/empty (possibly an omission).
-		if authDesc == "none" || authDesc == "" {
-			if tcpAddr, ok2 := ln.Addr().(*net.TCPAddr); ok2 && !tcpAddr.IP.IsLoopback() {
-				slog.Warn("bootstrap: listener bound to non-loopback address without auth; ensure network-level isolation",
-					slog.String("listener", ref.String()),
-					slog.String("addr", ln.Addr().String()),
-					slog.Bool("wildcard_bind", tcpAddr.IP.IsUnspecified()),
-					slog.Bool("explicit_auth_none", explicitAuthNone(cfg.authChain)))
-			}
-		}
-
-		servers = append(servers, boundServer{
-			name: ref.String(),
-			srv: &http.Server{
-				Handler:           rtr.Handler(),
-				ReadHeaderTimeout: 10 * time.Second,
-				ReadTimeout:       30 * time.Second,
-				WriteTimeout:      30 * time.Second,
-				IdleTimeout:       60 * time.Second,
-			},
-			ln:        ln,
-			owned:     owned,
-			shutGrace: cfg.shutGrace,
-			authDesc:  authDesc, // OPS-09: passed to phase7ServeAll startup log
-		})
-	}
-	return servers, nil
-}
-
-// closeOwnedSockets closes all sockets that bootstrap owns (i.e. not caller-injected).
-func closeOwnedSockets(servers []boundServer) {
-	for _, prev := range servers {
-		if prev.owned {
-			_ = prev.ln.Close()
-		}
-	}
-}
-
-// phase7ServeAll starts all servers in background goroutines and returns a channel
-// that receives errors and is closed when all servers have stopped.
-func (b *Bootstrap) phase7ServeAll(servers []boundServer) chan error {
-	n := len(servers)
-	httpErrCh := make(chan error, n)
-	pending := int32(n)
-	for _, bs := range servers {
-		bs := bs // capture
-		go func() {
-			slog.Info("bootstrap: HTTP server starting",
-				slog.String("listener", bs.name),
-				slog.String("addr", bs.ln.Addr().String()),
-				slog.String("auth", bs.authDesc))
-			err := bs.srv.Serve(bs.ln)
-			if err != nil && !errors.Is(err, http.ErrServerClosed) {
-				httpErrCh <- fmt.Errorf("%s listener: %w", bs.name, err)
-			}
-			if atomic.AddInt32(&pending, -1) == 0 {
-				close(httpErrCh)
-			}
-		}()
-	}
-	return httpErrCh
-}
-
-// resolveListener returns the net.Listener for a listenerConfig. When
-// cfg.net is set (caller-injected), it is returned directly (owned=false).
-// Otherwise a new TCP socket is bound to cfg.addr (owned=true).
-//
-// CORR-03: when cfg.tls is non-nil, the TCP listener is wrapped with
-// tls.NewListener so that TLS termination happens at the listener level.
-// The pre-bound caller-injected net listener (cfg.net) is returned as-is
-// since callers are expected to handle TLS on their own pre-bound socket.
-func resolveListener(cfg listenerConfig) (ln net.Listener, owned bool, err error) {
-	if cfg.net != nil {
-		return cfg.net, false, nil
-	}
-	tcpLn, listenErr := net.Listen("tcp", cfg.addr)
-	if listenErr != nil {
-		return nil, false, listenErr
-	}
-	if cfg.tls != nil {
-		return tls.NewListener(tcpLn, cfg.tls), true, nil
-	}
-	return tcpLn, true, nil
-}
-
-// shutdownTask represents a single server shutdown operation with its name and
-// grace period. Extracted from boundServer to allow tests to inject arbitrary
-// shutdown functions without requiring a real http.Server.
-type shutdownTask struct {
-	name      string
-	shutGrace time.Duration
-	shutdown  func(context.Context) error
-}
-
-// shutdownAllServers drains all servers in parallel. Each task uses a
-// per-server context derived from the parent ctx:
-//   - when shutGrace > 0 (set via WithListenerShutdownGrace), the parent ctx
-//     is wrapped with context.WithTimeout(ctx, shutGrace) so shutGrace is an
-//     upper bound *within* the global shutdownTimeout budget — never a parallel
-//     budget that can outlive global shutdown.
-//   - when shutGrace == 0, the shared ctx is passed through unchanged (server
-//     inherits the global shutdownTimeout deadline).
-//
-// Errors are aggregated via errors.Join so operators see every failure.
-//
-// OPS-01: each task carries a name so shutdown log lines identify the listener.
-// R2-03: parent ctx (not context.Background) is the timeout parent so the
-// global shutdownTimeout always bounds per-listener drains.
-func shutdownAllServers(ctx context.Context, tasks []shutdownTask) error {
-	slog.Info("bootstrap: draining HTTP servers")
-	type drainResult struct {
-		err error
-	}
-	resultCh := make(chan drainResult, len(tasks))
-	for _, task := range tasks {
-		task := task // capture
-		go func() {
-			shutCtx, cancel := shutdownCtxFor(ctx, task.shutGrace)
-			defer cancel()
-			err := task.shutdown(shutCtx)
-			if err != nil {
-				slog.Error("bootstrap: HTTP server drain failed",
-					slog.String("listener", task.name), slog.Any("error", err))
-				// OPS-01: wrap with the listener name so the returned error chain
-				// preserves the same attribution that already lives in the slog
-				// line. Without this, errors.Join collapses to opaque "shutdown
-				// failed" lines once the goroutine returns and operators can no
-				// longer tell which listener tripped from the error object alone.
-				err = fmt.Errorf("listener %q shutdown: %w", task.name, err)
-			} else {
-				slog.Info("bootstrap: HTTP server drained", slog.String("listener", task.name))
-			}
-			resultCh <- drainResult{err: err}
-		}()
-	}
-	var errs []error
-	for range tasks {
-		r := <-resultCh
-		if r.err != nil {
-			errs = append(errs, r.err)
-		}
-	}
-	return errors.Join(errs...)
-}
-
-// boundServersToTasks converts []boundServer into []shutdownTask for shutdownAllServers.
-func boundServersToTasks(servers []boundServer) []shutdownTask {
-	tasks := make([]shutdownTask, len(servers))
-	for i, bs := range servers {
-		tasks[i] = shutdownTask{
-			name:      bs.name,
-			shutGrace: bs.shutGrace,
-			shutdown:  bs.srv.Shutdown,
-		}
-	}
-	return tasks
 }
 
 // phase8StartWorkers starts the WorkerGroup using the caller-supplied runCtx
